+++ api/tests/team/integration/infrastructure/repositories/user-orga-settings-repository_test.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 
 import { UserOrgaSettingsCreationError } from '../../../../../src/shared/domain/errors.js';
-import { UserOrgaSettings } from '../../../../../src/shared/domain/models/UserOrgaSettings.js';
+import { UserOrgaSettings } from '../../../../../src/team/domain/models/UserOrgaSettings.js';
 import { userOrgaSettingsRepository } from '../../../../../src/team/infrastructure/repositories/user-orga-settings-repository.js';
 import { catchErr, databaseBuilder, expect, knex, sinon } from '../../../../test-helper.js';
