+++ api/tests/tooling/domain-builder/factory/build-user-orga-settings.js
@@ -1,6 +1,6 @@
 import { User } from '../../../../src/identity-access-management/domain/models/User.js';
 import { Organization } from '../../../../src/organizational-entities/domain/models/Organization.js';
-import { UserOrgaSettings } from '../../../../src/shared/domain/models/UserOrgaSettings.js';
+import { UserOrgaSettings } from '../../../../src/team/domain/models/UserOrgaSettings.js';
 
 /*
  * /!\ We can not use standard entity builders because of bidirectional relationships (a.k.a. cyclic dependencies)
